move collate to batching
