+++ examples/poyo/datasets/wrapper.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch
 
 from temporaldata import Data, Interval
-from torch_brain.data.collate import pad8
+from torch_brain.batching.collate import pad8
 from torch_brain.utils import isin_interval
 from torch_brain.dataset import DatasetIndex, Dataset
