+++ pyproject.toml
@@ -59,6 +59,7 @@ Documentation = "https://torch-brain.readthedocs.io/en/latest/"
 [tool.setuptools]
 packages = [
     "torch_brain",
+    "torch_brain.batching",
     "torch_brain.data",
     "torch_brain.dataset",
     "torch_brain.nn",
