+++ torch_brain/batching/__init__.py
@@ -1,0 +1,51 @@
+from .collate import (
+    chain,
+    collate,
+    pad,
+    pad8,
+    pad2d,
+    pad2d8,
+    track_batch,
+    track_mask,
+    track_mask8,
+    track_mask2d,
+    track_mask2d8,
+)
+
+__all__ = [
+    "chain",
+    "collate",
+    "pad",
+    "pad8",
+    "pad2d",
+    "pad2d8",
+    "track_batch",
+    "track_mask",
+    "track_mask8",
+    "track_mask2d",
+    "track_mask2d8",
+    "Dataset",
+]
+
+# see docs/source/api_reference.py
+__api_ref__ = {
+    "description": None,
+    "sections": [
+        {
+            "title": "torch_brain.batching.collate",
+            "autosummary": [
+                "chain",
+                "collate",
+                "pad",
+                "pad8",
+                "pad2d",
+                "pad2d8",
+                "track_batch",
+                "track_mask",
+                "track_mask8",
+                "track_mask2d",
+                "track_mask2d8",
+            ],
+        },
+    ],
+}
