+++ torch_brain/data/__init__.py
@@ -1,54 +1,13 @@
 from . import sampler
-from .collate import (
-    chain,
-    collate,
-    pad,
-    pad8,
-    pad2d,
-    pad2d8,
-    track_batch,
-    track_mask,
-    track_mask8,
-    track_mask2d,
-    track_mask2d8,
-)
+from . import collate
 from .dataset import Dataset
 
-__all__ = [
-    "chain",
-    "collate",
-    "pad",
-    "pad8",
-    "pad2d",
-    "pad2d8",
-    "track_batch",
-    "track_mask",
-    "track_mask8",
-    "track_mask2d",
-    "track_mask2d8",
-    "Dataset",
-]
+__all__ = ["Dataset"]
 
 # see docs/source/api_reference.py
 __api_ref__ = {
     "description": None,
     "sections": [
-        {
-            "title": "torch_brain.data.collate",
-            "autosummary": [
-                "chain",
-                "collate",
-                "pad",
-                "pad8",
-                "pad2d",
-                "pad2d8",
-                "track_batch",
-                "track_mask",
-                "track_mask8",
-                "track_mask2d",
-                "track_mask2d8",
-            ],
-        },
         {
             "title": "torch_brain.data.dataset [DEPRECATED]",
             "autosummary": [
